+++ apps/kyb-app/src/helpers/prepareInitialUIState.ts
@@ -1,16 +1,17 @@
 import { UIState } from '@/components/organisms/DynamicUI/hooks/useUIStateLogic/types';
 import { UIPage } from '@/domains/collection-flow';
 import { CollectionFlowContext } from '@/domains/collection-flow/types/flow-context.types';
-import { CollectionFlowManager } from '@ballerine/common';
+import { getCollectionFlowState } from '@ballerine/common';
 
 export const isPageCompleted = (page: UIPage, context: CollectionFlowContext) => {
-  const flow = new CollectionFlowManager(context);
-
-  const result = flow.state().isStepCompleted(page.stateName!);
+  const collectionFlow = getCollectionFlowState(context);
+  const isStepCompleted = collectionFlow?.steps?.find(
+    step => step.stepName === page.stateName,
+  )?.isCompleted;
 
   if (!page.stateName) return false;
 
-  return result;
+  return isStepCompleted;
 };
 
 export const prepareInitialUIState = (