+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.61
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.61
+
+## 0.6.60
+
+### Patch Changes
+
+- Change
+- Updated dependencies
+  - @ballerine/common@0.9.48
+  - @ballerine/workflow-core@0.6.60
+
+## 0.6.59
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.47
+  - @ballerine/workflow-core@0.6.59
+
 ## 0.6.58
 
 ### Patch Changes